Allow adding new entries with `edit` command

+++ cmd/edit.go
@@ -9,18 +9,24 @@ import (
 	"github.com/spf13/cobra"
 	"gitlab.com/JanMa/go-pass/pkg/git"
 	"gitlab.com/JanMa/go-pass/pkg/store"
+	"gitlab.com/JanMa/go-pass/pkg/store/entry"
 	"gitlab.com/JanMa/go-pass/pkg/util"
 )
 
 func editPassword(cmd *cobra.Command, args []string) {
-	entry, err := PasswordStore.FindEntry(args[0])
-	exitOnError(err)
-	exitOnError(entry.Decrypt())
+	e, exists := PasswordStore.FindEntry(args[0])
 	tmp, err := util.TmpFile()
 	exitOnError(err)
-	val, err := entry.Value()
-	exitOnError(err)
-	exitOnError(ioutil.WriteFile(tmp, []byte(val), 0600))
+	if exists == nil {
+		exitOnError(e.Decrypt())
+		val, err := e.Value()
+		exitOnError(err)
+		exitOnError(ioutil.WriteFile(tmp, []byte(val), 0600))
+	} else {
+		fmt.Println("Creating new entry", args[0])
+		e = entry.New(args[0], PasswordStore.Path+"/"+args[0]+".gpg")
+		PasswordStore.InsertEntry(e)
+	}
 	editor := exec.Command(getEditor(), tmp)
 	editor.Stdin = os.Stdin
 	editor.Stdout = os.Stdout
@@ -30,15 +36,15 @@ func editPassword(cmd *cobra.Command, args []string) {
 	newVal, err := ioutil.ReadFile(tmp)
 	exitOnError(err)
 	exitOnError(os.Remove(tmp))
-	if val == string(newVal) {
+	if val, _ := e.Value(); val == string(newVal) {
 		fmt.Println("Password unchanged")
 		return
 	}
-	entry.Insert(string(newVal))
-	iD := PasswordStore.FindGpgID(entry.Path)
+	e.Insert(string(newVal))
+	iD := PasswordStore.FindGpgID(e.Path)
 	recv, err := store.ParseGpgID(iD)
-	exitOnError(entry.Encrypt(recv))
-	git.AddFile(entry.Path, fmt.Sprintf("Edit %s with %s", args[0], getEditor()))
+	exitOnError(e.Encrypt(recv))
+	git.AddFile(e.Path, fmt.Sprintf("Edit %s with %s", args[0], getEditor()))
 }
 
 func getEditor() string {
